SCons: Allow to build Goost out of the same directory
Added root `SConstruct` which is a convenience/compatibility script
allowing to build Goost in the same way as building Godot Engine
so the usage should be the same.

Internally, this attempts to find Godot source and build the engine
with the Goost module, also providing customization such as
appending extra suffix to the resulting binaries and overriding
engine build name.

This also allows to build Goost with the existing built-in and community
`modules/`, which was pretty cumbersome to do (previously, you had to
manually append to the `custom_modules` option). But both approaches
are supported.