SCons: Allow `custom_modules`, `extra_suffix`, `profile` build options
These options were previously reserved to build this extension just like
the engine conveniently, but now it's possible to use these without
limitations.

A list of `custom_modules` is appended to the internal list of modules
to build from the command-line.

The default extra suffix (`goost`) can be extended (like `goost.custom`).

The `profile` option is not reserved to disabling modules in Goost now.
If set via the command-line, you can customize this with your own
configuration file.