SCons: Add engine custom patching system
The Godot core cannot be modified without tinkering with the engine
source, but in some cases, it's necessary to do so.

This change allows to circumvent the engine limitations and add small
enhancements to Godot before build more easily by applying custom
`git diff` patches.

This is disabled by default, and can be enabled with a new `use_godot_patches`
and `godot_patches` SCons build options:

```sh
scons use_godot_patches=yes godot_patches=misc/patches
```

The `godot_patches` build option always defaults to `misc/patches` value.
This way, only built-in patches will be applied, but any custom directory
path can be passed as well. If you use a custom directory, the built-in
patches will not be applied. It's recommended that you copy built-in
patches to your own directory instead.