Require instantiation of local instances for poly-related classes
Classes such as `PolyBoolean2D` can be used globally, but the default
parameters cannot be configured now. If you do need to configure
parameters, create a local instance of the same class with a newly
added `new_instance()` method.

This is done to prevent programming mistakes relating to
singletons usage, so that parameters from local operations cannot
affect the result of other operations throughout a project.

In most cases, it's not necessary to instantiate parameter classes,
since this is done per class now.