Rename `PolyMesh2D` to `PolyShape2D`
This may be confusing with CSG nodes in 3D, as mesh is used there
for one of the CSG shape generators.

We don't use `Shape` terminology here in `PolyNode2D` like for
`PolyCircle2D` or `PolyRectangle2D` as it's not necessarily about boolean
operators, but also representing an hierarchy of nested outlines, so
Goost reserves `Shape` terminology to lean closer to collision shapes.