Implement class dependency system
Declare a list of dependencies, so that it's not possible to compile the engine if a particular class is disabled if it's a dependency of other class.

Class should be registered via new `GOOST_REGISTER_CLASS` macro now, unless there are valid reasons to use `ClassDB::register_class` alone.

Individual classes can be currently disabled via `custom.py` which can be created at the root of the Goost source. For instance:
```python
# custom.py

goost_classes_disabled = [
    "LinkedList",
    "ListNode",
    "VariantResource",
    "VisualShape2D",
]
```

Note that currently this mostly affects run-time behavior. Classes may still be compiled, but this patch makes it easier to further optimize binaries for size in the future.