Do not compile sources associated with disabled Goost classes
Allows to optimize the resulting binary for size while compiling the engine, so it's not only about run-time behavior now.

Some hacks are removed, but some added. Using injection mechanism in `add_source_files` to override the default behavior so that we can filter out Goost-specific classes compiled in the module.

Some sources need to be renamed to match the `snake_case` version of class names, as the mechanism relies on `PascalCase` to `snake_case` string conversion.