Rename `Grid2D` to `VariantMap`
Mainly for avoiding confusion with UI elements. This name is also more explicit in describing the purpose of this class, and would be more similar with names such as `BitMap`.

Goost also has similar class named `VariantResource`, so again this makes the naming more consistent.

If ever needed, the 3D version could be named as `VariantGrid`, similarly to `GridMap` name for voxel stuff.