Get rid of "minor/major" terminology in `GridRect`
Users are free to interpret cell drawing as "minor" if they use divisions, but when there are no divisions, cells could be treated as "major" in the same manner, so just get rid of "minor/major" terms, because they can be ambiguous (the same applies to "primary/secondary" lines).

This change makes existing terminology of "cells/divisions" consistent throughout class.