Remove `main_script` property from `MixinScript`
Since there's no way to make arbitrary instances to use base types other than `Mixin`, there's not much benefit in disntinguishing between the "main" script and "mixins".

However, the first script added to `MixinScript` will be treated as the "main" one, in the sense that script editing is going to be automatically delegated to the first script.

This greatly simplifies both the core and the editor.