Use `config.py` to enable or disable GIF module
The previous way of disabling the module was a bit hacky. Since GIF module could be compiled independently of Goost, we still need a way to automatically disable the module if it's built as part of Goost where `goost_image_enabled=no` might be specified.

We rely on implicit module detection order. Since Goost is detected first, it's configuration is copied to base env in Godot, so other nested modules can refer to those options. If an option is not present, it means that the module is built independently of Goost.