Fix vertices generation order in `GoostGeometry2D.regular_polygon()`
Makes it "mathematically correct", the first vertex starts at `Vector2.RIGHT` direction. This also removes the need to generate points reverse to fix orientation.

The maximum number of points in `GoostGeometry2D.circle()` is unlimited now. The limit was hardcoded according to Godot's polygon vertex count limit for rendering, but it can be configured in fact.