Git plugin: Allow to customize path to repository
This adds `version_control/git/repository_path` project setting which allows to configure a path to existing Git repository.

This is useful in cases when you create prototypes as separate Godot projects, but wish to track changes in main project, or when repository is not initialized directly in Godot project.