Rename `draw_rect()` to `draw_region()` in `Debug2D`
The `draw_rect()` method used used to draw regions for the most part. This also makes it consistent with other methods in Goost such as `Random2D.point_in_region()`.

The method is renamed because of `draw_rectangle()` which is going to be implemented to draw rectangular shapes. Again, this would correspond to methods such as `GoostGeometry2D.rectangle()`.