Rename `VariantMap` to `Map2D`
This class was first renamed from `Grid2D` in #67 to avoid confusion with `GridRect`. After using the class more, it becomes apparent that `Map2D` better reflects the functionality here, because it allows to map 2D Cartesian coordinates to values.

This also avoids using `Variant` in a class name, which is frankly redundant.